Move Ticket, SwitchFs, Cnmt, Nro, Xci again
