+++ src/LibHac/Tools/Fs/Xci.cs
@@ -3,9 +3,8 @@ using LibHac.Common.Keys;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
 using LibHac.FsSystem;
-using LibHac.Tools.Fs;
 
-namespace LibHac;
+namespace LibHac.Tools.Fs;
 
 public class Xci
 {
